+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"embed"
 	"flag"
 	"io/fs"
@@ -117,8 +118,11 @@ func runMainMode(mailtoData *app.MailtoData, rawMailtoArg string) {
 		},
 		BackgroundColour: &options.RGBA{R: 27, G: 38, B: 54, A: 1},
 		OnStartup:        application.Startup,
-		OnShutdown:       application.Shutdown,
-		OnBeforeClose:    application.BeforeClose,
+		OnDomReady: func(ctx context.Context) {
+			platform.NotifyStartupComplete()
+		},
+		OnShutdown:    application.Shutdown,
+		OnBeforeClose: application.BeforeClose,
 		Bind: []interface{}{
 			application,
 			dummyComposerApp, // For binding generation